refactor(freeturn): вкладка в разделе «Туннели» + декомпозиция и визуальная полировка
- FreeTurn — вкладка на главной рядом с Sing-box (паттерн SingboxTunnelsTabSection),
  top-level пункт меню убран; /freeturn — redirect на /?tab=freeturn
- страница 679 строк → компоненты lib/components/freeturn/: FreeTurnTab
  (оркестратор: API/поллинг/состояние), ClientPanel, ServerPanel,
  ProcessCard (общий каркас клиент/сервер), ProcessStatsRow
- stat-плитки сверху и всегда видимы (статус+uptime, PID, режим, обфускация)
- самодельные select/checkbox → Dropdown/FormToggle из UI-кита
- лог процесса — свёрнутый details внизу карточки (не выталкивает форму);
  ошибка последнего запуска остаётся сверху
- импорт по ссылке — компактный блок под карточкой клиента
- бейджи «вкл» на вкладках Клиент/Сервер при работающем процессе
