+++ frontend/src/lib/components/freeturn/index.ts
@@ -1,0 +1,5 @@
+export { default as FreeTurnTab } from './FreeTurnTab.svelte';
+export { default as ClientPanel } from './ClientPanel.svelte';
+export { default as ServerPanel } from './ServerPanel.svelte';
+export { default as ProcessCard } from './ProcessCard.svelte';
+export { default as ProcessStatsRow } from './ProcessStatsRow.svelte';
