+++ frontend/src/lib/components/freeturn/uptime.ts
@@ -1,0 +1,9 @@
+/** Аптайм процесса для шапки карточки и stat-плитки: «N мин» / «Xч Yм». */
+export function formatUptime(startedAt?: string): string {
+	if (!startedAt) return '';
+	const ms = Date.now() - new Date(startedAt).getTime();
+	const mins = Math.floor(ms / 60000);
+	if (mins < 60) return `${mins} мин`;
+	const hrs = Math.floor(mins / 60);
+	return `${hrs}ч ${mins % 60}м`;
+}
